# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Giggle Grove API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BookType = typeof BookType[keyof typeof BookType];


export const BookType = {
  storybook: 'storybook',
  coloring_book: 'coloring_book',
} as const;

export interface Book {
  id: number;
  title: string;
  type: BookType;
  coverUrl: string;
  /** @nullable */
  previewUrl: string | null;
  /** @nullable */
  canvaLink?: string | null;
  /** @nullable */
  driveLink?: string | null;
  /** @nullable */
  instagramLink?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  ageRange?: string | null;
  createdAt?: string;
}

export type BookInputType = typeof BookInputType[keyof typeof BookInputType];


export const BookInputType = {
  storybook: 'storybook',
  coloring_book: 'coloring_book',
} as const;

export interface BookInput {
  title: string;
  type: BookInputType;
  coverUrl: string;
  previewUrl?: string;
  canvaLink?: string;
  driveLink?: string;
  instagramLink?: string;
  description?: string;
  ageRange?: string;
}

export interface BookUpdate {
  title?: string;
  coverUrl?: string;
  previewUrl?: string;
  canvaLink?: string;
  driveLink?: string;
  instagramLink?: string;
  description?: string;
  ageRange?: string;
}

export interface Influencer {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  instagramHandle?: string | null;
  referralCode: string;
  clicks: number;
  commissionTotal: number;
  createdAt?: string;
}

export interface InfluencerInput {
  name: string;
  email: string;
  instagramHandle?: string;
  location?: string;
}

export interface User {
  id: number;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  referredBy?: string | null;
  freePreviewsUsed?: number;
  createdAt: string;
}

export interface UserInput {
  email: string;
  name?: string;
  referralCode?: string;
}

export interface AdminCredentials {
  username: string;
  password: string;
}

export interface AdminSession {
  token: string;
  username: string;
}

export interface AdminSettings {
  freePreviewLimit: number;
}

export interface PreviewLimitUpdate {
  freePreviewLimit: number;
}

export interface StatsSummary {
  totalUsers: number;
  totalBooks: number;
  totalReferralClicks: number;
  activeInfluencers: number;
}