# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Giggle Grove API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: books
    description: Sample book operations
  - name: influencers
    description: Influencer operations
  - name: admin
    description: Admin operations
  - name: users
    description: User operations
  - name: stats
    description: Stats and summaries

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /books:
    get:
      operationId: listBooks
      tags: [books]
      summary: List all sample books
      responses:
        "200":
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
    post:
      operationId: createBook
      tags: [admin]
      summary: Upload/create a sample book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"

  /books/{id}:
    get:
      operationId: getBook
      tags: [books]
      summary: Get a single book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
    patch:
      operationId: updateBook
      tags: [admin]
      summary: Update a book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
    delete:
      operationId: deleteBook
      tags: [admin]
      summary: Delete a book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /influencers:
    get:
      operationId: listInfluencers
      tags: [influencers]
      summary: List all influencers
      responses:
        "200":
          description: List of influencers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Influencer"
    post:
      operationId: registerInfluencer
      tags: [influencers]
      summary: Register as influencer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InfluencerInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Influencer"

  /influencers/{id}/track-click:
    post:
      operationId: trackReferralClick
      tags: [influencers]
      summary: Track a referral link click
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Click tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Influencer"

  /users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users (admin)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Register/create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Admin token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Unauthorized

  /admin/preview-limit:
    patch:
      operationId: updatePreviewLimit
      tags: [admin]
      summary: Update the number of free previews allowed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreviewLimitUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"

  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get current admin settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"

  /stats/summary:
    get:
      operationId: getStatsSummary
      tags: [stats]
      summary: Dashboard summary — total users, books, referral clicks, active influencers
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsSummary"

  /stats/top-influencers:
    get:
      operationId: getTopInfluencers
      tags: [stats]
      summary: Top influencers by clicks and commissions
      responses:
        "200":
          description: Top influencers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Influencer"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Book:
      type: object
      required: [id, title, type, coverUrl, previewUrl]
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [storybook, coloring_book]
        coverUrl:
          type: string
        previewUrl:
          type: ["string", "null"]
        canvaLink:
          type: ["string", "null"]
        driveLink:
          type: ["string", "null"]
        instagramLink:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        ageRange:
          type: ["string", "null"]
        createdAt:
          type: string

    BookInput:
      type: object
      required: [title, type, coverUrl]
      properties:
        title:
          type: string
        type:
          type: string
          enum: [storybook, coloring_book]
        coverUrl:
          type: string
        previewUrl:
          type: string
        canvaLink:
          type: string
        driveLink:
          type: string
        instagramLink:
          type: string
        description:
          type: string
        ageRange:
          type: string

    BookUpdate:
      type: object
      properties:
        title:
          type: string
        coverUrl:
          type: string
        previewUrl:
          type: string
        canvaLink:
          type: string
        driveLink:
          type: string
        instagramLink:
          type: string
        description:
          type: string
        ageRange:
          type: string

    Influencer:
      type: object
      required: [id, name, email, referralCode, clicks, commissionTotal]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        instagramHandle:
          type: ["string", "null"]
        referralCode:
          type: string
        clicks:
          type: integer
        commissionTotal:
          type: number
        createdAt:
          type: string

    InfluencerInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
        instagramHandle:
          type: string
        location:
          type: string

    User:
      type: object
      required: [id, email, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: ["string", "null"]
        referredBy:
          type: ["string", "null"]
        freePreviewsUsed:
          type: integer
        createdAt:
          type: string

    UserInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
        name:
          type: string
        referralCode:
          type: string

    AdminCredentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    AdminSession:
      type: object
      required: [token, username]
      properties:
        token:
          type: string
        username:
          type: string

    AdminSettings:
      type: object
      required: [freePreviewLimit]
      properties:
        freePreviewLimit:
          type: integer

    PreviewLimitUpdate:
      type: object
      required: [freePreviewLimit]
      properties:
        freePreviewLimit:
          type: integer

    StatsSummary:
      type: object
      required: [totalUsers, totalBooks, totalReferralClicks, activeInfluencers]
      properties:
        totalUsers:
          type: integer
        totalBooks:
          type: integer
        totalReferralClicks:
          type: integer
        activeInfluencers:
          type: integer