# lib/api-zod/src/generated/types/book.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Giggle Grove API
 * OpenAPI spec version: 0.1.0
 */
import type { BookType } from './bookType';

export interface Book {
  id: number;
  title: string;
  type: BookType;
  coverUrl: string;
  /** @nullable */
  previewUrl: string | null;
  /** @nullable */
  canvaLink?: string | null;
  /** @nullable */
  driveLink?: string | null;
  /** @nullable */
  instagramLink?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  ageRange?: string | null;
  createdAt?: string;
}