# lib/api-zod/src/generated/types/bookInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Giggle Grove API
 * OpenAPI spec version: 0.1.0
 */
import type { BookInputType } from './bookInputType';

export interface BookInput {
  title: string;
  type: BookInputType;
  coverUrl: string;
  previewUrl?: string;
  canvaLink?: string;
  driveLink?: string;
  instagramLink?: string;
  description?: string;
  ageRange?: string;
}