# lib/api-zod/src/generated/types/influencer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Giggle Grove API
 * OpenAPI spec version: 0.1.0
 */

export interface Influencer {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  instagramHandle?: string | null;
  referralCode: string;
  clicks: number;
  commissionTotal: number;
  createdAt?: string;
}