# lib/api-zod/src/generated/types/influencerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Giggle Grove API
 * OpenAPI spec version: 0.1.0
 */

export interface InfluencerInput {
  name: string;
  email: string;
  instagramHandle?: string;
  location?: string;
}