# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Giggle Grove API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalUsers: number;
  totalBooks: number;
  totalReferralClicks: number;
  activeInfluencers: number;
}