# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Giggle Grove API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  referredBy?: string | null;
  freePreviewsUsed?: number;
  createdAt: string;
}