# lib/db/src/schema/books.ts
import { pgTable, text, serial, integer, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const booksTable = pgTable("books", {
  id: serial("id").primaryKey(),
  title: text("title").notNull(),
  type: text("type").notNull(),
  coverUrl: text("cover_url").notNull(),
  previewUrl: text("preview_url"),
  canvaLink: text("canva_link"),
  driveLink: text("drive_link"),
  instagramLink: text("instagram_link"),
  description: text("description"),
  ageRange: text("age_range"),
  createdAt: timestamp("created_at").defaultNow().notNull(),
});

export const insertBookSchema = createInsertSchema(booksTable).omit({ id: true, createdAt: true });
export type InsertBook = z.infer<typeof insertBookSchema>;
export type Book = typeof booksTable.$inferSelect;